+++ src/wifi/model/non-ht/ofdm-phy.cc
@@ -650,7 +650,7 @@ OfdmPhy::IsAllowed(const WifiTxVector& /*txVector*/)
 uint32_t
 OfdmPhy::GetMaxPsduSize() const
 {
-    return 4095;
+    return WIFI_PSDU_MAX_LENGTH;
 }
 
 MHz_u